Added some casting for the IDE

+++ src/packet/packet.c
@@ -131,7 +131,7 @@ pkt_status_code pkt_decode(const char *data, const size_t len, pkt_t *pkt) {
         return statusCode;
     }
     // TODO calculate CRC1 with tr set to 0 !?
-    uint32_t crc1Calculated = crc32(crc32(0L, Z_NULL, 0), (Bytef *) data+0, (uInt) sizeof(char)*8);
+    uint32_t crc1Calculated = (uint32_t) crc32(crc32(0L, Z_NULL, 0), (Bytef *) data + 0, (uInt) sizeof(char) * 8);
     if(pkt_get_crc1(pkt) != crc1Calculated) {
         return E_CRC;
     }
@@ -188,7 +188,7 @@ pkt_status_code pkt_encode(pkt_t *pkt, char *buf, size_t *len) {
     memcpy(buf+4, &pkt_timestamp, sizeof(uint32_t));
 
     // TODO calculate CRC1 with tr set to 0 !?
-    uint32_t pktn_crc1 = htonl(crc32(crc32(0L, Z_NULL, 0), (Bytef *) buf+0, (uInt) sizeof(char)*8));
+    uint32_t pktn_crc1 = htonl((uint32_t) crc32(crc32(0L, Z_NULL, 0), (Bytef *) buf + 0, (uInt) sizeof(char) * 8));
     //uint32_t pktn_crc1 = htonl(pkt_get_crc1(pkt));
     statusCode = pkt_set_crc1(pkt, ntohl(pktn_crc1));
     if(statusCode != PKT_OK) {
@@ -219,7 +219,7 @@ pkt_status_code pkt_encode(pkt_t *pkt, char *buf, size_t *len) {
     /*
      * Ecriture du CRC2 dans le buffer
      */
-    uint32_t pktn_crc2 = htonl(crc32(crc32(0L, Z_NULL, 0), (Bytef *) buf+12, (uInt) pkt_get_length(pkt)));
+    uint32_t pktn_crc2 = htonl((uint32_t) crc32(crc32(0L, Z_NULL, 0), (Bytef *) buf + 12, (uInt) pkt_get_length(pkt)));
     //uint32_t pktn_crc2 = htonl(pkt_get_crc2(pkt));
     statusCode = pkt_set_crc2(pkt, ntohl(pktn_crc2));
     if(statusCode != PKT_OK) {
